modulize fetch API
