+++ configuration-app-sample/src/public/js/helpers/fetchApi.js
@@ -1,0 +1,19 @@
+import { getBearerToken } from "./getBearerToken.js";
+
+const token = getBearerToken();
+
+export async function fetchApi(url, method) {
+  try {
+    let response1 = await fetch(url, {
+      method: method,
+      headers: {
+        Authorization: `Bearer ${token}`
+      }
+    });
+    let response2 = await response1.json();
+    return response2;
+  } catch (error) {
+    console.log(error);
+    // location.href = `${host}/index.html`;
+  }
+}
